iknowyou-cloud/ai-server/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
sqlalchemy==2.0.23
pymysql==1.1.0
pinecone-client==2.2.4
openai==1.3.7
langchain==0.0.350
langchain-openai==0.0.2
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.2
transformers==4.35.2
torch==2.1.1
requests==2.31.0
aiofiles==23.2.1
python-multipart==0.0.6 
